improved: --oracle-path
1. replace --oracle-sqlpath with --oracle-path
2. sqlpath is not stable

ORACLE_PATH
Environment variable to specify the location of SQL scripts. If SQL*Plus cannot find the file in ORACLE_PATH, or if ORACLE_PATH is not set, it searches for the file in the current working directory.
Not applicable to Windows

SQLPATH
Environment variable or Windows registry entry to specify the location of SQL scripts. SQL*Plus searches for SQL scripts, including login.sql, in the directories specified by SQLPATH, and in the subdirectories of SQLPATH directories. SQLPATH is a colon-separated list of directories. There is no default value set in UNIX installations.
In Windows, SQLPATH is defined in a registry entry during installation. For more information about the SQLPATH registry